rename data/index=>common
