+++ src/components/Viewer/common/ImageDownloader.tsx
@@ -3,7 +3,7 @@ import JSZip from "jszip"
 import FileSaver from "file-saver"
 
 import { escape } from "utils"
-import { allSolidNames } from "data"
+import { allSolidNames } from "data/common"
 
 /**
  * Utility class to download image thumbnails. Do NOT use in production
